chore: refactor api spec transform / utils
